clean the code and organise the folders
