+++ portofolio/src/components/shared/ImageReveal/ImageReveal.scss
@@ -18,8 +18,4 @@
         right: 0;
         background-color: white;
     }
-    // img{
-    //     //width: 600;
-    //     //position: absolute;
-    // }
 }
