+++ portofolio/src/index.css
@@ -19,6 +19,7 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
 /* slide used in gitHub button in Navbar*/
 .slide {
   --color: #00CD8F;
