+++ portofolio/tailwind.config.js
@@ -7,10 +7,8 @@ module.exports = {
     extend: {
       colors: {
         'primary-green': '#00CD8F',
-        'primary-gray': '#EEEEEE',
-        'secondary-gray':'#595a5e',
-        'tertiary-gray':'#D3D6E0',
-        'secondary-blue' : '#1D1B27'
+        'secondary-gray': '#EEEEEE',
+        'secondary-blue' : '#17181C'
       }
     },
   },
